AI - ML/Semester 2/ecolens/requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
tensorflow-hub==0.16.0
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
opencv-python-headless==4.9.0.80
plotly==5.19.0
requests==2.31.0
tqdm==4.66.2
python-dotenv==1.0.1
secure-smtplib==0.1.1
altair==5.2.0
